+++ shoop/admin/__init__.py
@@ -46,6 +46,7 @@ class ShoopAdminAppConfig(AppConfig):
             "shoop.admin.modules.services.forms:WaivingCostBehaviorComponentForm",
             "shoop.admin.modules.services.forms:WeightLimitsBehaviorComponentForm",
             "shoop.admin.modules.services.forms:GroupAvailabilityBehaviorComponentForm",
+            "shoop.admin.modules.services.forms:RoundingBehaviorComponentForm",
         ],
         "service_behavior_component_form_part": [
             "shoop.admin.modules.services.weight_based_pricing.WeightBasedPricingFormPart"
@@ -58,6 +59,10 @@ class ShoopAdminAppConfig(AppConfig):
     }
 
     def ready(self):
+        from shoop.core.order_creator.signals import order_creator_finished
+        from shoop.admin.modules.orders.receivers import handle_custom_payment_return_requests
+
+        order_creator_finished.connect(handle_custom_payment_return_requests, dispatch_uid='shoop.admin.order_create')
         validate_templates_configuration()